hiberman: Fix crashes in cat subcommand
The 'hiberman cat' subcommand used for development and testing was
crashing in two places:

First, when specifying the --log switch, it was crashing because the I/O
size coming down to the BufferedDiskFile was not aligned. This because
read_to_end() starts with a 32-byte buffer and expands it, which doesn't
meet the page size requirement of DiskFile. Also, read_to_end() is not
even what we want, since that reads the entire file, rather than reading
until the first nul character. Instead, do what the log replay code
does, and make a BufReader to do read_until(0).

Second, the variant without the --log switch was also crashing, because
copy() obviously reads until it sees EOF. DiskFile considered it a
criminal offense to read EOF. Usually it would be, as it indicates our
file size was too small for actual use. But in cat's case this isn't
true, so remove the assert, and downgrade the error!() to a warn!()
(since we still want to know if we hit EOF outside of the cat command).

BUG=b:243982794
TEST=hiberman cat [--log] resume_log, observe no crash

Change-Id: I0d9d5771cff36457f4eec4d9d7dae19f01cddc81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3860389
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>