cryptohome: Make UserSession non-ref-counted
The UserSession objects don't have to be ref-counted. Ref-counting makes
ownership and lifetime story vague, and also means that the logout code
can't really enforce that the active sessions are shut down.

This commit changes the class to be non-ref-counted, and updates all
callsites to use either raw pointers or std::unique_ptr.

BUG=b:243846478
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: Ie034a04efcfa220c7282c3dbaad483dd1ef0748c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3859848
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>