power_manager: Increase the resuspend time for dark resume from existing 5 secs to 10 secs.
When we connect USB4 dock with display to the DUT in S0ix, the device wakes from S0ix and resuspend back in 5 secs.5 secs is too less of time for complete thunderbolt enumeration and display detection behind that.
We need to give enough time before resuspending so that all the TBT devices are enumerated completely before resuspend in triggered.

BUG=b:239790615
TEST= Put the device into S0ix with dark resume enabled; connect a USB4 dock with display to the DUT, while DUT is in S0ix. On connecting the USB4, expect the DUT to do a full wake

Change-Id: I6a6c09d65973846b05c6e5621bf7613384187dcd
Signed-off-by: Abhijeet Rao <abhijeet.rao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3864292
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Abhijeet Rao <abhijeet.rao@intel.corp-partner.google.com>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Abhijeet Rao <abhijeet.rao@intel.corp-partner.google.com>