secagentd: build: stop using board bpftool
Use the build host's bpftool instead of using the board bpftool. The
board is not guaranteed to be the same arch as the build host; as such
invoking board bpftool when the archs are mismatched will fail.

Instead use the build host's bpftools to generate vmlinux.h and the
skeletons.

BUG=b:244222378
TEST=Compiled across different architectures
(amd64-generic,arm64-generic, arm64) and verify that BPFs build just
fine.

Change-Id: Ieffabe3e6a1fb8bf871fb448455640d778d7a6e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3868383
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Commit-Queue: Jason Ling <jasonling@chromium.org>
Tested-by: Aashay Shringarpure <aashay@google.com>