attestation: add hardcoded workaround for early DT EK certs
The first batch of Dauntless chips was produced with misconfigured EK
certificates, in particular the 'Issuer' field name is of a wrong
format.

The attestation code uses certificate issuer name value as a key for
matching the CA public key used to validate the EK.

This patch adds code which would detect the misconfigured field name,
and check if it matches the value from the EK certificates of the
Dauntless batch in question. In case of the match the value is
replaced with the hardcoded expected value of "CROS D2 CIK"

BUG=b:243565780
TEST=verified that the following sequence succeeds on Chromeos devices
     with both kinds of EK certificate formats, misconfigured and
     correct.

  $ tpm-manager initialize
  $ tpm-manager verify_endorsement

Change-Id: If003c423f14245da5f2c1199c2f28ae777095f58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3864503
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>