shill: passpoint: fix Passpoint credentials validation
This patch fixes a validation bug with Passpoint credentials where
credentials would be rejected incorrectly if only a domain suffix match
list is passed for verifying the server certificate with the system
trusted CAs.

It is important to correctly support this case because this is how
Android populates an Android PasspointConfiguration object from a
WifiConfiguration object on Android P. This allows ARC to create
Passpoint credentials as close as possible to Android.

BUG=b:244105387
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Id8b4c048c50e64fa44cc4fa49690ce4c4c113802
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3864643
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>