camera: use EGL platform display on VMs
When running on VM, initializing the default EGL display opens the
primary device and would take DRM master if Chrome hasn't already done
so, which blocks Chrome from starting because the Chrome GPU process
needs to be the DRM master.

The issue is easily reproducible when running VM tests, which emits
large amounts of `restart ui` in a short period of time. When Chrome is
blocked all the VM tests would fail.

BUG=b:242140143
TEST=Run VM tests with amd64-generic image in QEMU.
TEST=Run image_processor_test on device and in VM.

Change-Id: Ief701f4a5122059dd0dfe5941c6522d57be10404
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3869965
Tested-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>