cryptohome: Refactor reply callback style in Add operations
Pass a callback with error status paraemter to
AuthSession::AddAuthFactor() and AuthSession::AddCredentials()
to populate the error value. This simplifies the code from passing
the reply callback to the lower levels in the stack.
As we are changing the unittests for add operations this CL also
changes the touched test functions to use TestFuture.

BUG=b:240596931
TEST=tast run DUT cryptohome* hwsec*

Change-Id: I13eabfdb2e704abf5566b3c25a47e51bc7f4252b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3863000
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Yi Chou <yich@google.com>