vm_tools: concierge: Rewrite GetCpuQuota()
Now that platform2's feature library supports parameters, we no
longer have to define lots of features as a workaround. This CL
replaces the features with one new feature with a parameter.

This is safe to do because the Finch experiment that uses the old
parameter has already ended. We'll start a new experiment with
the new feature name.

This CL depends on crrev.com/c/3864460 (Chromium)

BUG=b:205762950
TEST=Add --enable-features=CrOSLateBootArcVmInitialThrottle:quota/42
  to /etc/chrome_dev.conf, reboot DUT, sign in, wait for ARCVM to
  boot, verify /sys/fs/cgroup/cpu/arcvm/cpu.cfs_quota_us is 42000.

Change-Id: I6fca792d0bee6466fc089f07a53795de5dc5a4d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3867913
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>