shill: Add state in Network
Add a `State state_;` member variable in the Network. It can be one of
the three state: idle, configuring, and connected, which indicates the
L3 connectivity state of the Network.

The new state has the following mapping with the current code:
- state is connected == there is a Connection object and Connection has
  been set up;
- state is not connected == there is no Connection object;
- state is idle == has_started;
- state is not idle == !has_started.
Thus this patch should not change any logic in the code.

Unit tests for the state will be added in the follow-up patches.

BUG=b:232177767
TEST=FEATURES="test" emerge-${BOARD} shill

Change-Id: I1a037db29668b650fe22f4ff946ca86d7ec685e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3864754
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>