hps: only dump valid registers in DumpHpsRegisters()
Instead of blindly dumping every register from 0 to N, and assuming they
all exist and are readable, check from 0 to 127 against a list of known
registers.

In order to do so, introduce a table of register info structs (RegInfo)
and a helper function to look up a given register number in the table.
In future commits we can expand the table to indicate how each register
value should be decoded.

BUG=b:227109909
TEST=FEATURES=test emerge-brya hpsd hps-tool

Change-Id: If0243b857e42ae17df4b803396b9f1d418dca436
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3864016
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>