init: let factory mode imply fast
We call clobber-state with `factory` flag in two places:
(1) Factory wiping [1], which is the last step of the factory process.
(2) Factory reset [2], which is used to reset the device back to
    shipping mode.
And in this CL, we would like to enable `fast` if `factory` flag is on.

For (1), it has already been reviewed by the security team in [3] and a
CL for fast flag was drafted in [4].
For (2), in our current implementation, we already called
`clobber-state` with fast flag in [2].

Ref:
[1] src/platform/factory/py/gooftool/wipe.py
[2] src/platform/factory_installer/factory_reset.sh
[3] b/204836990
[4] https://crrev.com/c/3510920

BUG=b:229168577
TEST=CQ

Change-Id: Ibd28fb262c3e3579a7304889d08b844f6635751d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3864602
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Stimim Chen <stimim@chromium.org>
Tested-by: Phoebe Wang <phoebewang@chromium.org>
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>