diagnostics: Put mojo service into another class
Put all the mojo service into another class. The new class is
responsible for initialization and lifecycle management. This is a part
of mojo service manager migration.
Also add forward declaration of the mojo interfaces in this class so we
don't need to add mojo dependency to all the build target in BUILD.gn.
Note that mojo::Remote can not be forward declared because it request
the type defined in the mojo interface class.

BUG=b:244407986
TEST=FEATURES=test emerge-zork diagnostics

Change-Id: Ib785697176820e47500af4d86039d63091f72cd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3867331
Reviewed-by: Kerker Yang <kerker@chromium.org>
Tested-by: Chung-Sheng Wu <chungsheng@google.com>
Commit-Queue: Chung-Sheng Wu <chungsheng@google.com>