camera: split frame annotator to a shared library
Since the size of Skia might be too big, we decided to split the whole
frame annotator to a separate library that could be loaded when
possible.

This CL put the original frame annotator to libs/ folder and add a
loader stream manipulator for loading the real frame annotator when
possible.

BUG=b:236679365
TEST=`USE='camera_feature_face_detection camera_feature_frame_annotator' emerge-dedede cros-camera-libs`

Change-Id: I1e467ecd7ceff5932bb114dbb704c8205777e7db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3864195
Commit-Queue: Ren-Pei Zeng <kamesan@chromium.org>
Tested-by: Tommy Chiang <ototot@google.com>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>