featured: Add Cpuidle TEO/Menu experiment
These can be changed at runtime on most kernels. (On some kernels, the
file will be /sys/devices/system/cpu/cpuidle/current_governor_ro, which
is read-only.)

NB: the default is "menu" on some platforms, and "teo" on others, so we
include an experiment for each. One of these will be redundant with the
default (non-experiment).

Carve out selinux policy for /sys/devices/system/cpu/cpuidle/*, to allow
featured to write to it, and to still allow regular daemons to read.

Align the JSON file mostly [*] according to 'cros format'.

[*] Ignoring the sorting: http://shortn/_NPjzjtYpGE
    It definitely makes sense to keep "name:" fields first.

BUG=b:240453755
TEST=add enable-features=CrOSLateBootCpuidleMenu to /etc/chrome_dev.conf

Change-Id: I0de9b9a18a4c5d0cbd7aa62afa2bc7a8532e6740
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3864899
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>