cryptohome: Use Intents for switching to appropriate AuthSession status
Refactor the AuthSession class to use a set of authorized AuthIntents
when switching into a new state after an auth factor authentication.

This a non-functional change - it's a preparation for implementing
the lightweight authentication, which will only authorize the
"VERIFY_ONLY" intent and not the powerful "DECRYPT" one.

BUG=b:240596931
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: I22c65925b72129b54fa351e24f792c4115152f5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3867655
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>