cryptohome: Implement lightweight AuthSession auth
Add a lightweight authentication support into the AuthenticateAuthFactor
D-Bus method.

This special mode of authentication is faster and uses an
in-memory credential verifier. This code path is only used when the
AUTH_INTENT_VERIFY_ONLY was passed in StartAuthSessionRequest, which
will be done by Chrome in unlock-only scenarios (the Lock Screen or
visiting a web password manager).

BUG=b:240596931
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: I885fc74ad95a4166db173540176d59156d59d891
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3867656
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>