diagnostics: Executor delegate accepts mojo invitation from executor
In this CL, delegate daemon accepts the mojo invitation from executor
and bind the receiver to the DelegateImpl. From now on, it's ready to
process remote's request.

One thing to note, timing to terminate the daemon:
I'd like to make this daemon more flexible, so the remote is able to
launch it once and run several requests in a row, hence we shouldn't
terminate it once we fulfill one request. Instead, remote SHOULD take
care of the life cycle of the daemon, that is, in this CL, remote
SHOULD disconnect the mojo connection so the daemon will stop.

BUG=b:243603240
TEST=Test with following CLs

Change-Id: Ifa7acb85e4674ed9bb4560dbba0772436ad5123f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3877838
Reviewed-by: Chung-Sheng Wu <chungsheng@google.com>
Tested-by: Kerker Yang <kerker@chromium.org>
Commit-Queue: Kerker Yang <kerker@chromium.org>