cryptohome: delete objects on the correct thread
The CL ensures that objects that are created on the mount thread are
also queued for deletion on that thread. Without this, when the program
exits and deletes these objects on the main (non-sequenced) thread,
objects that do sequence checking (like Timer::~Timer) will crash.

These sorts of crashes can fill up /var/spool/crash on the stateful
volume, which will then make a tast suite fail.

BUG=b:244751769
TEST=FEATURES=test emerge-amd64-generic cryptohome
              tast run <VM> hwsec.CrossVersionLogin
	      # check stateful /var/spool/crash for files

Change-Id: I0fc4901eef3c180f2eb58e1ed957ce193b393c93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3874211
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>