rmad: Always poll WP status in WpDisablePhysical state
|RunState| is called when starting a state, either at boot up or
transitioning from a previous state. A special case is that when rmad
starts, it attempts to do a state transition using the current state,
and if the transition succeeds, the original state's |RunState| is not
called.

Hence We don't need to disable the polling timer even when we're ready
to transition to the next state. In the case of transitioning from a
previous state, the timer starts and triggers a reboot once HWWP is
disabled. On the next boot up, we transition to the next state before
calling |RunState|.

BUG=b:245249442
TEST=FEATURES=test emerge-octopus rmad

Change-Id: Iee9dd01e600f9eba2f50ba8e07a5750d62adbadb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3875942
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Gene Chang <genechang@google.com>