rmad: Record EC reboot status in WpDisablePhysical state
If the repair tech shuts down the device when disconnecting the battery,
the state will auto-transition to WpDisableComplete state on the next
boot, but in fact we need an additional EC reboot here to disable EC WP
in the future. Hence we store an additional variable to indicate that we
have done an EC reboot ourselves.

BUG=b:245267127
TEST=FEATURES=test emerge-octopus rmad

Change-Id: I00e9d974ae6dbb3474f244d5428eeb33f28c37b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3875946
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Gene Chang <genechang@google.com>