system_api: Clarify timestamp units in `DeviceSetupCertificateMetadata`
This proto is shared verbatim with Google - and its presubmit is not OK
with a scalar `uint64` field to denote a timestamp since its unit (e.g.
seconds vs milliseconds) is not obvious from the name / type of the
field. Ref. - go/proto-best-practices-checkers#time-unit

BUG=b:240398262
TEST=CQ

Change-Id: Ie13abbf463210ae84722e49488e4fc08b5b957bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3873614
Tested-by: Kush Sinha <sinhak@chromium.org>
Commit-Queue: Kush Sinha <sinhak@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>