cryptohome: Test AuthSession verifier state
Add unit tests for the state of the credential verifier in the
AuthSession class. The tests ensure that the verifier is populated in
basic scenarios.

Follow-up commits will fix a few operations that forgot to populate the
verifier and cover those with regression tests in a similar fashion.

BUG=b:245238313
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: I24abbaaefc1311e75118aa3869a842439e1bb21a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3879503
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>