trunks: Pass null expiration param correctly
In pinweaver_client selftest, one of the InsertLeaf command incorrectly
passed a non-null expiration delay (0, which means roughly the same as
nullopt), but that will fail when protocol version is < 2 because of the
sanity check in trunks. Fix that function call to pass nullopt
correctly.

BUG=b:245609478
TEST=pinweaver selftest

Change-Id: I81b2964b375aa973692be7409c3a858587db2458
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3881302
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>