Reland "init: move stateful mount code to C++"
This is a reland of commit f84436944455282458370d0d96731f73a2740e20,
with the addition of a fix for properly setting the
LogicalVolumeManager unique pointer. This was causing errors on
guybrush and brya. Also no longer removing the setting of STATE_DEV
and DEV_IMAGE for lvm devices to make sure they are properly set
when used later in chromeos_startup.sh.

Original change's description:
> init: move stateful mount code to C++
>
> This CL moves the mount stateful code to c++.
>
> Move Resume from hibernate the LVM way to C++
>
> Following chromium:3700916 fix hibernate resume in the C++ code.
>
> BUG=b:172210918
> TEST=build_image --board=volteer
>
> Cq-Depend: chromium:3854499
> Change-Id: I4e8f160038311865dbf09146ad7c0286255df980
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2999418
> Commit-Queue: Alexis Savery <asavery@chromium.org>
> Reviewed-by: Mike Frysinger <vapier@chromium.org>
> Tested-by: Alexis Savery <asavery@chromium.org>

BUG=b:172210918
TEST=build_image --board=brya

Change-Id: I3f690d0c5dc1f7109003e3f2f8e2f281efd8383b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3932138
Tested-by: Alexis Savery <asavery@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>