cryptohome: Store AuthFactorType in verifier
Expose the AuthFactorType from the CredentialVerifier class. Pass the
factor type (when we know it, i.e., in non-legacy APIs) to the place
that creates a verifier, and don't create a verifier for the factors
for which it's unsupported.

This change will allow implementing the proper support of ephemeral
users, for which there's no VaultKeyset or AuthFactor stored physically
on the disk, and the only information about the configured factor (if
there's any) is stored in the UserSession and the corresponding
CredentialVerifier.

BUG=b:240596931
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: Ife4e119e228df9d4cc9673fb9d85c591cded73b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3882301
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>