patchpanel: ndproxy: RA only trigger RouterDiscovery once
In previous implementation of ndproxy, when an RA is received, ADD_ADDR
event is triggered on each downstream interface corresponding to that
uplink. This patch changes that into a single ROUTER_DETECTED event with
the uplink id, and it's up to GuestIPv6Service to find corresponding
downlinks. This enables future migration to other mechanism of handling
this event.

No end-to-end behavior change in this patch.

BUG=b:228585272
TEST=`FEATURES=test emerge-atlas patchpanel`
TEST=`tast run local arc.IPv6Connectivity`
TEST=manual: switch wifi/eth and observe address/route set correctly

Change-Id: If670ff4c574097182bf3e2d593c32a1fc9a57eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3886360
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Taoyu Li <taoyl@chromium.org>