cryptohome: Ephemeral AddAuthFactor sets verifier
Create a credential verifier when AddAuthFactor is called for an
ephemeral user.

This will allow implementing the lock screen support for ephemeral users
with a password - for this to work we'll need to be able to check the
credential against a verifier in AuthenticateAuthFactor.

BUG=b:244754956,b:240596931
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: Ibc631ffd99077a7a3f5bcfe639ef8ad8a6474c80
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3886436
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>