hiberman: Add LoginToResumeReady metric
One important question we'll want an answer to is how long users are
actually waiting for a hibernate resume. Resume runs as the user is
authenticating, so the existing metrics that show total resume time
don't equate 1:1 with perceived delay, since the user spends some
or all of that time interacting with authentication (eg typing in
their password).

Add a metric that starts when the d-bus call from Chrome is received
(which happens just after login is known to be successful) and ends when
we are done with most disk I/O and about ready to attempt the resume.

BUG=b:232184089
TEST=Hibernate and resume

Change-Id: I9c5c5eb2d66a03be18f455750103b5168d2256b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3887838
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>