power: Ignore user activity during resume from hibernate
Powerd has a flavor of suspend request called RESUME_FROM_DISK_PREPARE.
This reuses the suspend callback machinery to prepare the system to
resume from hibernate. Since this is not a normal suspend, we don't want
to abort this request if user activity comes in, since we're in the
middle of resuming from hibernate.

HandleEventInResumingFromHibernate() already correctly ignores most
events when the state is RESUMING_FROM_HIBERNATE. However if the powerd
state machine hasn't fully transitioned there due to pending suspend
callbacks, the state might be WAITING_FOR_NORMAL_SUSPEND_DELAYS. If we
are here and receive a user activity event, we end up aborting the
RESUME_FROM_DISK_PREPARE. This is not what we want, and hiberman gets
confused waiting for a completion that never arrives.

Ignore user activity events in the state
WAITING_FOR_NORMAL_SUSPEND_DELAYS if the SuspendRequest flavor is
RESUME_FROM_DISK_PREPARE.

BUG=b:245995197
TEST=FEATURES=test emerge-brya-hibernate powerd

Change-Id: Idf7bc14593c22eed7ad0522ca58f84c58f9b8780
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3888204
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>