cryptohome: Store label in credential verifier
Update all authentication code paths to store the key/factor label in
the CredentialVerifier.

The label has been used by legacy APIs for skipping the lightweight
authentication when a different key is supplied. It used to be stored in
the KeyData proto in UserSession. As we're implementing the new
AuthFactor-based API to have feature parity, we need to do similar
checks, but without involving the legacy KeyData protos. Hence the CL
updates the code to store the label as a separate string field; we also
put it into CredentialVerifier as it's closely related to the
credential and the AuthFactorType.

BUG=b:244754956,b:240596931
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: I455e725e86e93df23b482e3fcc22e67dc5c19170
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3888880
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>