crash: Loosen coreboot banner string match a little
Since CB:61308, coreboot is writing loglevel markers at the start of
each line in the CBMEM console log. In /sys/firmware/log which is read
by crash-reporter, these currently show up as a single non-printable
byte. In the future, we may also adapt the kernel driver to translate
them on the fly into a readable prefix (e.g. "[DEBUG]  ". In order to
make the crash-reporter robust to all these variations, just accept any
extra prefix at the beginning of the line for the banner string.

Note that the Pre-CBMEM console overflow banner is special because it is
printed directly in the CBMEM console driver and bypasses the normal
coreboot console stack. Therefore, it cannot contain any loglevel info.

BUG=b:244707357
TEST=kernel_collector_test

Change-Id: I54fae8162a757f6419c4160281d4d89fa48d9694
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3892045
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Julius Werner <jwerner@chromium.org>
Tested-by: Julius Werner <jwerner@chromium.org>