shill: vpn: Check pid file before starting charon
At present, for an L2TP/IPsec service, an immediate reconnect attempt after the disconnection may fail. It is because, when the second connection starts, the charon process from the previous connection may not be fully stopped, and thus IPsecConnection fails to start a second charon process. This issue will occur if you set "auto reconnect" to true, and the default network is changed.

This patch rewrites the flow of making a new IPsec connection, to make sure the old charon process disappeared by checking that it is not included in the list of the current terminating processes in ProcessManager before the new charon process starts.

BUG=b:241857231
TEST=unit: FEATURES="test" emerge-$BOARD shill
TEST=tast: tast run $DUT network.VPNConnect.l2tp_ipsec*

Change-Id: Ia28db684047352f5e5e96625bb13584f63f92722
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3893879
Tested-by: Arisa Tsuji <arisai@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Arisa Tsuji <arisai@google.com>