vm_tools: cros_im: Move wl_seat logic to the backend
Currently cros_im takes a wl_seat from the frontend, but this will only
work for applications already running under Wayland and for X11 apps
it'd be easier for the backend to create it. This CL updates the backend
to always create a single new wl_seat and use that for all requests
instead of requiring the frontend to provide one. This is fine as seats
in Exo are all basically the same and we don't support any sort of
'multi-seat' behaviour.

Also try and make things a bit more rigorous against racing with
instantiating wayland globals, so we only allow creating text_input
objects once all required globals are ready.

BUG=b:232048280
TEST=../test/run_tests.py, run gedit with cros_im and input JP text.

Change-Id: Ia13fcba06ff7835f527ed2848351cedd0d02a0d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3890592
Reviewed-by: Nic Hollingum <hollingum@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Reviewed-by: Sophia Lin <sophialin@google.com>