shill: Establish core lifecycle for WiFiPhy objects
Initial implementation for a WiFiPhy class in Shill. This CL establishes
the basic life-cycle of a WiFiPhy object, and introduces unit test to
validate that life-cycle.

The WiFiPhy lifecycle should map 1:1 with the lifecycle of an actual phy
on the device. For example, if at any time `iw list` shows that there is
a phy on the device, then there should be a WiFiPhy object associated
with that Phy in Shill. This is done by tracking netlink messages that
indicate that a phy has been created or destroyed.

This CL does not implement any of the actual responsibilities of the
WiFiPhy class. Additional information available in the doc:
go/shill-wifiphy-thoughts

BUG=b:243701309
TEST=FEATURES=test emerge-hatch shill-net shill

Change-Id: Ic41bdb7a246b26e9a0bc16f87a715bc5f52a5cb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3897770
Reviewed-by: David Ruth <druth@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>