cryptohome: Expose the credential verifier from UserSession
Rather than having separate Has and Get functions, I just replaced the
Has with Get that returns null on "not available". If in the future we
more to multiple verifiers then I think Get can be more naturally
translated into Find, and Has stops making much sense, so I think this
is a reasonable direction.

The future intent is to make use of this from userdataauth for getting
factor information (which the verifier can provide) for ephemeral users.

BUG=b:245988742
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I7306417a162d91d4c6c5134a4d71e2f4b5b171dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3892196
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>