cryptohome: Add auth factor metadata to the credential verifiers
In line with the label and type info, also report metadata. Right now I
implemented this with a bit of a cheat: I just construct the metadata
directly in the verifier implementation. Because we only have one
implementation, and it's a password factor, and the password metadata is
trivial, we can get away with this.

The more general solution would require either capturing the factor
itself, or at least its metadata, in the verifier factory (the same way
we currently pass through label). However, this is tricky to do today
because of all the non-AuthFactor codepaths that use verifier, which
would have a hard time constructing a metadata to pass in. It should be
easier to implement if some of those old paths go away; but either way
we can defer implementing that until we need more general support for
other types of verifiers.

BUG=b:245988742
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I22475e5abeb26ad520a9cb742a706bfaf4db5630
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3892005
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>