cryptohome: Loosen up the persistent user existence check
The current check in ListAuthFactors for if a users exists (and is
persistent) is just based on whether or not they have a persistent
vault. But that's not actually good enough, if you create a new
persistent user you should be able to list their factors even if you
haven't set up a vault yet.

All this change does is loosen the check up a little, to do the same
thing as we do for ephemeral users and also check for an active
UserSession.

Before this change, if you had a setup flow of:
 - StartAuthSession (ephemeral=false)
 - CreatePersistentUser
 - PreparePersistentVault
Then you could only successfully call ListAuthFactors after the
CreatePersistentUser call, since only at that point would the have a
persistent directory. With this change, you can now call
ListAuthFactors immediately after the StartAuthSession call.

BUG=b:246586203
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ib480bd5524f648f39d94a73383fdec4e0cf958e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3894396
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>