typecd: Add PowerProfile class
Introduce a PowerProfile class, which represents the set of Source and
Sink Capabilities for a Type-C peripheral. These are derived from
reading the "usb_power_delivery" directory for a Type-C peripheral.

The PowerProfile object is currently just an empty class, but future
patches will expand this to include the lists of PDOs (Power Delivery
Objects) that the peripheral advertises.

Also, add support to create a PowerProfile object for a Partner.

BUG=b:245608929
TEST=- All unit tests pass.
     - Tested on brya connected to a Thunderbolt3 Dell WD19TB dock.
       PowerProfile object is created successfully (inspected via logs).

Change-Id: Ie30442dd65b23e12ca9feb32c4b79024b15306c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3894484
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Won Chung <wonchung@google.com>
Tested-by: Prashant Malani <pmalani@chromium.org>