cryptohome: Remove the duplicated AddAuthFactor code
Both AddAuthFactorViaUSS() and AddAuthFactorViaVaultKeyset() perform
the same AuthBlock operations. Remove the duplicate function calls
and merge the AuthBlock paths for AddauthFactor with USS and with
VK. This CL doesn't add any functional changes, but prepare the
AddAuthFactor to create a back up VK, when the USS is enabled.

BUG=b:246656102
TEST=tast run DUT cryptohome.* hwsec.*

Change-Id: I57d9a9e8b8aeae6130842fa25982fbd43ddc2de4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3895297
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>