libbrillo: Split stream WaitForData into read and write
Drop the support for READ_WRITE in |WaitForData| and
|WaitForDataBlocking| by splitting the functions into read and write
versions. This commit keeps the original implementation so the added
codes (either the read or the write version) are just a copy of the
existing ones.

Supporting READ_WRITE makes the functions confusing since the callers
may misunderstand the criteria of the callback being invoked. Also,
READ_WRITE is never passed as the argument by any caller.

The new function |WaitForDataRead/Write| follow the same style of other
functions like |CanRead/Write| and |Read/WriteAsync|.

BUG=b:241931908
TEST=USE="-cros-debug" FEATURES=test emerge-volteer libbrillo

Change-Id: I2a3b8aba1e051b30aaa32d37dd379c8c595e1d9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3891346
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Tested-by: Wei-Luan Wang <weiluanwang@google.com>