cryptohome: Run ExtendAuthSession when AuthSession is authenticated
Earlier we started the timer as soon as AuthSession started, but since
the behavior has changed. We now start the timer when AuthSession is
authenticated. Sadly, this was missed earlier. Changing it now as I was
working around this area already.

BUG=b:230945496
TEST=FEATURES=test emerge-zork cryptohome

Change-Id: I5e4a935d693b450312306f5e9dbada3810825f21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3901492
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>